+++ tests/ui/components/Section.js
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Section = (props) => {
 	return (
 		<div {...props} style={{display: 'inline-block'}} />